Time: 67 ms (28.34%), Space: 109.2 MB (43.96%) - LeetHub

+++ 0853-car-fleet/0853-car-fleet.cpp
@@ -13,7 +13,7 @@ public:
         stack<double> st;
 
         for (int i = 0; i < pair.size(); i++) {
-            double cur = (target - pair[i].first) / pair[i].second;
+            double cur = (double)(target - pair[i].first) / pair[i].second;
 
             if (st.empty()) {
                 st.push(cur);
